==> 01 - Getting Bearer Token.ipynb <==
import requests

public_key = 'DABED781-3155-46C5-8297-02852684B959'
private_key = 'B154EE91-4892-4FA9-9ED1-D335E7D08065'

# This will retrieve our BEARER_TOKEN

headers = {'application': 'x-www-form-urlencoded'}

data = 'grant_type=client_credentials&client_id='+ public_key + '&client_secret=' + private_key

response = requests.post('https://api.tcgplayer.com/token', headers=headers, data=data)

response.json()

# Bearer Token - This need's to be in the headers of the query

BEARER_TOKEN = "LtdVEcwDtrDBD-Y6vsHg9yRhN6BSZWRPBdGRso47EtG1IDuKs5hKdIrYA8xBJhROoXjjAjKP4Ct6bbQa3XUdarnWVzrm0Yt-ybcdmxyh0A8N1mgQpy4Vo8VEFwaeX8uWCF2XolI2C9_RLklpAIX5agMN-x9mIPyY6t8Bkd33Yk9nYWxBHIOWZNJmE-WSyMjJ1JVmcTBfX0XDIV8BUefCiQez0-lghzR4swiiE5FE_WVuc_xacCPw_hSRVs5_IhFgg6aCOY9ACO4c9IwahJDHFOo0Q3nZTbsnbZxpCf4FNbVFKlYuP0kBUAee0RcwvWctpout6w"
