==> 02 - List of Every Card.ipynb <==
list_of_set_names = ['Limited+Edition+Alpha',
 'Limited+Edition+Beta',
 'Unlimited+Edition',
 'Arabian+Nights',
 'Antiquities',
 'Revised+Edition',
 'Legends',
 'The+Dark',
 'Fallen+Empires',
 'Fourth+Edition',
 'Ice+Age',
 'Chronicles',
 'Renaissance',
 'Homelands',
 'Alliances',
 'Mirage',
 'Visions',
 'Fifth+Edition',
 'Portal',
 'Weatherlight',
 'Tempest',
 'Stronghold',
 'Exodus',
 'Portal+Second+Age',
 'Unglued',
 'Urza%27s+Saga',
 'Anthologies',
 'Urza%27s+Legacy',
 'Sixth+Edition',
 'Urza%27s+Destiny',
 'Portal+Three+Kingdoms',
 'Starter+1999',
 'Mercadian+Masques',
 'Battle+Royale',
 'Nemesis',
 'Starter+2000',
 'Prophecy',
 'Invasion',
 'Beatdown',
 'Planeshift',
 'Seventh+Edition',
 'Apocalypse',
 'Odyssey',
 'Deckmasters+2001',
 'Torment',
 'Judgment',
 'Onslaught',
 'Legions',
 'Scourge',
 'Eighth+Edition',
 'Mirrodin',
 'Darksteel',
 'Fifth+Dawn',
 'Champions+of+Kamigawa',
 'Unhinged',
 'Betrayers+of+Kamigawa',
 'Saviors+of+Kamigawa',
 'Ninth+Edition',
 'Salvat+2005',
 'Ravnica%3a+City+of+Guilds',
 'Guildpact',
 'Dissension',
 'Coldsnap',
 'Time+Spiral',
 'Planar+Chaos',
 'Future+Sight',
 'Tenth+Edition',
 'Masters+Edition',
 'Lorwyn',
 'Duel+Decks%3a+Elves+vs.+Goblins',
 'Morningtide',
 'Shadowmoor',
 'Eventide',
 'From+the+Vault%3a+Dragons',
 'Masters+Edition+II',
 'Shards+of+Alara',
 'Duel+Decks%3a+Jace+vs.+Chandra',
 'Conflux',
 'Duel+Decks%3a+Divine+vs.+Demonic',
 'Alara+Reborn',
 'Magic+2010',
 'Commander+Theme+Decks',
 'From+the+Vault%3a+Exiled',
 'Planechase',
 'Masters+Edition+III',
 'Zendikar',
 'Duel+Decks%3a+Garruk+vs.+Liliana',
 'Premium+Deck+Series%3a+Slivers',
 'Worldwake',
 'Duel+Decks%3a+Phyrexia+vs.+The+Coalition',
 'Rise+of+the+Eldrazi',
 'Deck+Builder%27s+Toolkit',
 'Duels+of+the+Planeswalkers',
 'Archenemy',
 'Magic+2011',
 'From+the+Vault%3a+Relics',
 'Duel+Decks%3a+Elspeth+vs.+Tezzeret',
 'Scars+of+Mirrodin',
 'Magic+Online+Deck+Series',
 'Premium+Deck+Series%3a+Fire+&+Lightning',
 'Momir+Basic+Event+Deck',
 'Salvat+2011',
 'Masters+Edition+IV',
 'Mirrodin+Besieged',
 'Deck+Builder%27s+Toolkit+2011',
 'Duel+Decks%3a+Knights+vs.+Dragons',
 'New+Phyrexia',
 'Commander',
 'Magic+2012',
 'From+the+Vault%3a+Legends',
 'Duel+Decks%3a+Ajani+vs.+Nicol+Bolas',
 'Innistrad',
 'Premium+Deck+Series%3a+Graveborn',
 'Dark+Ascension',
 'Duel+Decks%3a+Venser+vs.+Koth',
 'Avacyn+Restored',
 'Planechase+2012',
 'Magic+2013',
 'From+the+Vault%3a+Realms',
 'Duel+Decks%3a+Izzet+vs.+Golgari',
 'Return+to+Ravnica',
 'Commander%27s+Arsenal',
 'Duel+Decks%3a+Mirrodin+Pure+vs.+New+Phyrexia',
 'Gatecrash',
 'Duel+Decks%3a+Sorin+vs.+Tibalt',
 'Dragon%27s+Maze',
 'Modern+Masters',
 'Magic+2014',
 'From+the+Vault%3a+Twenty',
 'Duel+Decks%3a+Heroes+vs.+Monsters',
 'Theros',
 'Commander+2013',
 'Born+of+the+Gods',
 'Duel+Decks%3a+Jace+vs.+Vraska',
 'Journey+into+Nyx',
 'Modern+Event+Deck',
 'Conspiracy',
 'Vintage+Masters',
 'Magic+2015',
 'From+the+Vault%3a+Annihilation',
 'Duel+Decks%3a+Speed+vs.+Cunning',
 'Khans+of+Tarkir',
 'Commander+2014',
 'Duel+Decks+Anthology',
 'Fate+Reforged',
 'Duel+Decks%3a+Elspeth+vs.+Kiora',
 'Dragons+of+Tarkir',
 'Tempest+Remastered',
 'Modern+Masters+2015',
 'Magic+Origins',
 'From+the+Vault%3a+Angels',
 'Duel+Decks%3a+Zendikar+vs.+Eldrazi',
 'Battle+for+Zendikar',
 'Zendikar+Expeditions',
 'Commander+2015',
 'Legendary+Cube',
 'Oath+of+the+Gatewatch',
 'Duel+Decks%3a+Blessed+vs.+Cursed',
 'Welcome+Deck+2016',
 'Shadows+over+Innistrad',
 'Eternal+Masters',
 'Eldritch+Moon',
 'From+the+Vault%3a+Lore',
 'Conspiracy%3a+Take+the+Crown',
 'Duel+Decks%3a+Nissa+vs.+Ob+Nixilis',
 'Kaladesh',
 'Kaladesh+Inventions',
 'Treasure+Chests',
 'Commander+2016',
 'You+Make+the+Cube',
 'Planechase+Anthology',
 'Aether+Revolt',
 'Modern+Masters+2017',
 'Duel+Decks%3a+Mind+vs.+Might',
 'Welcome+Deck+2017',
 'Amonkhet',
 'Amonkhet+Invocations',
 'Commander+Anthology',
 'Archenemy%3a+Nicol+Bolas',
 'Hour+of+Devastation',
 'Commander+2017',
 'Ixalan',
 'Duel+Decks%3a+Merfolk+vs.+Goblins',
 'Iconic+Masters',
 'Explorers+of+Ixalan',
 'From+the+Vault%3a+Transform',
 'Unstable',
 'Rivals+of+Ixalan',
 'Masters+25',
 'Duel+Decks%3a+Elves+vs.+Inventors',
 'Challenger+Decks',
 'Dominaria',
 'Commander+Anthology+Volume+II',
 'Battlebond',
 'Signature+Spellbook%3a+Jace',
 'Global+Series%3a+Jiang+Yanggu+&+Mu+Yanling',
 'Core+Set+2019',
 'Commander+2018',
 'Guilds+of+Ravnica+Mythic+Edition',
 'Guilds+of+Ravnica',
 'Spellslinger+Starter+Kit',
 'Guilds+of+Ravnica+Guild+Kits',
 'Game+Night',
 'Ultimate+Masters',
 'Ravnica+Allegiance+Mythic+Edition',
 'Ravnica+Allegiance',
 'Ravnica+Allegiance+Guild+Kits',
 'Challenger+Decks+2019',
 'War+of+the+Spark+Mythic+Edition',
 'War+of+the+Spark',
 'Modern+Horizons',
 'Signature+Spellbook%3a+Gideon',
 'Core+Set+2020',
 'Commander+2019']

!pip install selenium

from selenium import webdriver
from selenium.webdriver.common.by import By
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
from selenium.common.exceptions import TimeoutException

# example option: add 'incognito' command line arg to options
option = webdriver.ChromeOptions()
#option.add_argument("--incognito")

# create new instance of chrome in incognito mode
driver = webdriver.Chrome(executable_path='C:/Users/adaml/Downloads/chromedriver_win32/chromedriver.exe')

# go to website of interest
driver.get("https://gatherer.wizards.com/Pages/Search/Default.aspx?action=advanced&cmc=+>=[0]")

# get all of the titles for the financial values
titles_element = driver.find_elements_by_xpath("//span[@class='cardTitle']")
titles = [x.text for x in titles_element]

titles


# YouTube Video: https://www.youtube.com/watch?v=Z3vFdtZ7d-g
from selenium import webdriver
import pandas as pd
import time

start = time.time()

MAX_PAGE_NUM = 196

driver = webdriver.Chrome(executable_path='C:/Users/adaml/Downloads/chromedriver_win32/chromedriver.exe')

names = []

for i in range(0, MAX_PAGE_NUM + 1):

    url = "http://gatherer.wizards.com/Pages/Search/Default.aspx?page=" + str(i) + "&output=standard&action=advanced&cmc=+%3E=[0]"
    driver.get(url)

    names_scrape = driver.find_elements_by_xpath("//span[@class='cardTitle']")
    names_lst = [x.text for x in names_scrape]
    names.extend(names_lst)
    

magic_df = pd.DataFrame(
        {'Name': names})


driver.close()

end = time.time()
print (end - start)

len(names)

magic_df.head()

magic_df.to_csv('cardnames.csv')
