==> 03 - List of Every SKU.ipynb <==
# Get the list of every card name

import pandas as pd

card_names_df = pd.read_csv('cardnames.csv')
card_names_df.head()

card_names = list(card_names_df['Name'])

# Create a dataframe of all the SKUs and save it to csv

import requests
import time

startTime = time.time()
#Set up API
BEARER_TOKEN = "LtdVEcwDtrDBD-Y6vsHg9yRhN6BSZWRPBdGRso47EtG1IDuKs5hKdIrYA8xBJhROoXjjAjKP4Ct6bbQa3XUdarnWVzrm0Yt-ybcdmxyh0A8N1mgQpy4Vo8VEFwaeX8uWCF2XolI2C9_RLklpAIX5agMN-x9mIPyY6t8Bkd33Yk9nYWxBHIOWZNJmE-WSyMjJ1JVmcTBfX0XDIV8BUefCiQez0-lghzR4swiiE5FE_WVuc_xacCPw_hSRVs5_IhFgg6aCOY9ACO4c9IwahJDHFOo0Q3nZTbsnbZxpCf4FNbVFKlYuP0kBUAee0RcwvWctpout6w"

headers = {'Accept': 'application/json', 'Authorization': 'bearer ' + BEARER_TOKEN}


#Query
cardNumber = 1
cards = []
intervalTime = time.time()
url = "http://api.tcgplayer.com/v1.27.0/catalog/products"

for card in card_names:
    # set up a timer for every 100 cards
    if cardNumber % 100 == 0:
        print("Elapsed Time for Last 100 Cards: ", time.time() - intervalTime)
        intervalTime = time.time()
    
    # loop every card name into here
    querystring = {"categoryId":"1","productName":card,"includeSkus":"true"}

    response = requests.request("GET", url, params=querystring, headers = headers)

    json_object = response.json()

    for i in range(len(json_object.get('results'))):
        #Sku Data Fetch
        for sku in json_object.get('results')[i].get('skus'):
            cards.append(sku['skuId'])
    cardNumber += 1

print(cards)
print('Elapsed Time: ',time.time()-startTime)

df = pd.DataFrame(cards)

df.to_csv('SKUs.csv')


df.shape
