==> 04 - Fetch all prices for each SKU.ipynb <==
# Import SKU numbers

import pandas as pd

SKU_df = pd.read_csv('SKUs.csv')
SKU_df.head()

SKU_list = list(SKU_df['0'])
SKU_list = sorted(SKU_list)
print (len(SKU_list))
print (SKU_list)

SKU_list

len(SKU_list)

# Setup the data frame

columns = {'sku':[], 'lowPrice':[], 'lowestShipping':[], 'lowestListingPrice':[], 'marketPrice':[], 'directLowPrice':[]}

price_df = pd.DataFrame(data = columns)

price_df.head()

# Just one SKU

import requests
import json
import pandas as pd
import time

startTime = time.time()
#Set up API
BEARER_TOKEN = "LtdVEcwDtrDBD-Y6vsHg9yRhN6BSZWRPBdGRso47EtG1IDuKs5hKdIrYA8xBJhROoXjjAjKP4Ct6bbQa3XUdarnWVzrm0Yt-ybcdmxyh0A8N1mgQpy4Vo8VEFwaeX8uWCF2XolI2C9_RLklpAIX5agMN-x9mIPyY6t8Bkd33Yk9nYWxBHIOWZNJmE-WSyMjJ1JVmcTBfX0XDIV8BUefCiQez0-lghzR4swiiE5FE_WVuc_xacCPw_hSRVs5_IhFgg6aCOY9ACO4c9IwahJDHFOo0Q3nZTbsnbZxpCf4FNbVFKlYuP0kBUAee0RcwvWctpout6w"

headers = {'Accept': 'application/json', 'Authorization': 'bearer ' + BEARER_TOKEN}


sku_number = 2231
url = "http://api.tcgplayer.com/v1.27.0/pricing/sku/" + str(sku_number)
response = requests.request("GET", url, headers = headers)
json_object = response.json()

low_price = json_object.get('results')[0].get('lowPrice')
lowest_shipping = json_object.get('results')[0].get('lowestShipping')
lowest_listing_price = json_object.get('results')[0].get('lowestListingPrice')
market_price = json_object.get('results')[0].get('marketPrice')
direct_low_price = json_object.get('results')[0].get('directLowPrice')

price_df.loc[0] = [int(sku_number), low_price, lowest_shipping, lowest_listing_price, market_price, direct_low_price]
price_df

# Loop through all SKUs

count = 0
intervalTime = time.time()
start = time.time()

for SKU in SKU_list:
    if count % 1000 == 0:
        print("Elapsed Time for Last 1000 SKU: ", time.time() - intervalTime)
        intervalTime = time.time()
        print (count)

    url = "http://api.tcgplayer.com/v1.27.0/pricing/sku/" + str(SKU)
    response = requests.request("GET", url, headers = headers)
    json_object = response.json()

    low_price = json_object.get('results')[0].get('lowPrice')
    lowest_shipping = json_object.get('results')[0].get('lowestShipping')
    lowest_listing_price = json_object.get('results')[0].get('lowestListingPrice')
    market_price = json_object.get('results')[0].get('marketPrice')
    direct_low_price = json_object.get('results')[0].get('directLowPrice')
    
    price_df.loc[count] = [SKU, low_price, lowest_shipping, lowest_listing_price, market_price, direct_low_price]
    
    count += 1
    
end = time.time()
print (end - start)

prices_df.to_csv('skus_with_prices.csv')
