==> 05 - Removing Unknown Prices.ipynb <==
The purpose of this code is to remove all the rows that have no pricing data. It's done by:

   * Importing a table of values
   * Create a new column ('drop_zeros') that's the sum of the prices for that row
   * Then drop all rows with a 'drop_zeros' value of 0
        * This means there was no prices in that row
   * Remove the 'drop_zeros' column
   * Set the index to SKU

import pandas as pd

csv_file = 'Data/sku_prices_071519.csv'

df = pd.read_csv(csv_file)

col_list = ['lowPrice', 'lowestShipping', 'lowestListingPrice', 'marketPrice', 'directLowPrice']

df['drop_zeros'] = df[col_list].sum(axis = 1)

df = df[df.drop_zeros != 0]

df = df.drop('drop_zeros', axis = 1)

df.set_index('sku', inplace = True)

df.to_csv('Data/cleaned_' + csv_file[-10:])
